+++ monix-reactive/shared/src/main/scala/monix/reactive/subjects/PublishSubject.scala
@@ -17,12 +17,13 @@
 
 package monix.reactive.subjects
 
-import monix.execution.Ack.{Stop, Continue}
+import monix.execution.Ack.{Continue, Stop}
+import monix.execution.atomic.Atomic
+import monix.execution.atomic.PaddingStrategy.LeftRight128
 import monix.execution.{Ack, Cancelable}
 import monix.reactive.internal.util.PromiseCounter
 import monix.reactive.observers.Subscriber
 import monix.reactive.subjects.PublishSubject.State
-import monix.execution.atomic.Atomic
 import scala.annotation.tailrec
 import scala.concurrent.Future
 import scala.util.control.NonFatal
@@ -41,7 +42,7 @@ final class PublishSubject[T] private () extends Subject[T,T] { self =>
    * NOTE: the stored vector value can be null and if it is, then
    * that means our subject has been terminated.
    */
-  private[this] val stateRef = Atomic(State[T]())
+  private[this] val stateRef = Atomic.withPadding(State[T](), LeftRight128)
 
   private
   def onSubscribeCompleted(subscriber: Subscriber[T], ex: Throwable): Cancelable = {
@@ -54,12 +55,12 @@ final class PublishSubject[T] private () extends Subject[T,T] { self =>
     stateRef.get.subscribers.size
 
   /*
-     * NOTE: onSubscribe is in contention with onNext, onComplete and onError,
-     * thus access to the subscribers/isDone state is done through CAS on an
-     * Atomic and if the new subscriber must be fed an initial set of events,
-     * then we enforce a happens-before relationship by means of the
-     * FreezeOnFirstOnNextSubscriber (the purpose of calling onSubscribeContinue)
-     */
+   * NOTE: onSubscribe is in contention with onNext, onComplete and onError,
+   * thus access to the subscribers/isDone state is done through CAS on an
+   * Atomic and if the new subscriber must be fed an initial set of events,
+   * then we enforce a happens-before relationship by means of the
+   * FreezeOnFirstOnNextSubscriber (the purpose of calling onSubscribeContinue)
+   */
   @tailrec
   def unsafeSubscribeFn(subscriber: Subscriber[T]): Cancelable = {
     val state = stateRef.get
@@ -83,77 +84,89 @@ final class PublishSubject[T] private () extends Subject[T,T] { self =>
 
   def onNext(elem: T): Future[Ack] = {
     val state = stateRef.get
-    // at some point we are going to notice the most recent subscribers
-    val subscribers = state.subscribers
+    val subscribersArray = state.cache
 
-    if (subscribers eq null) Stop else {
-      val iterator = subscribers.iterator
-      // counter that's only used when we go async, hence the null
-      var result: PromiseCounter[Continue.type] = null
-
-      while (iterator.hasNext) {
-        val subscriber = iterator.next()
-        // using the scheduler defined by each subscriber
-        import subscriber.scheduler
-
-        val ack = try subscriber.onNext(elem) catch {
-          case NonFatal(ex) => Future.failed(ex)
-        }
-
-        // if execution is synchronous, takes the fast-path
-        if (ack.isCompleted) {
-          // subscriber canceled or triggered an error? then remove
-          if (ack != Continue && ack.value.get != Continue.AsSuccess)
-            unsubscribe(subscriber)
-        }
-        else {
-          // going async, so we've got to count active futures for final Ack
-          // the counter starts from 1 because zero implies isCompleted
-          if (result == null) result = PromiseCounter(Continue, 1)
-          result.acquire()
-
-          ack.onComplete {
-            case Continue.AsSuccess =>
-              result.countdown()
-
-            case _ =>
-              // subscriber canceled or triggered an error? then remove
-              unsubscribe(subscriber)
-              result.countdown()
-          }
-        }
+    if (subscribersArray eq null) {
+      val set = state.subscribers
+      if (set == null) Stop else {
+        val update = state.refresh
+        // If CAS fails, it means we have new subscribers;
+        // not bothering to recreate the cache for now
+        stateRef.compareAndSet(state, update)
+        sendOnNextToAll(update.cache, elem)
       }
-
-      // has fast-path for completely synchronous invocation
-      if (result == null) Continue else {
-        result.countdown()
-        result.future
-      }
+    }
+    else {
+      sendOnNextToAll(subscribersArray, elem)
     }
   }
 
-  def onError(ex: Throwable): Unit = {
-    onCompleteOrError(ex)
-  }
+  def onError(ex: Throwable): Unit =
+    sendOnCompleteOrError(ex)
+  def onComplete(): Unit =
+    sendOnCompleteOrError(null)
 
-  def onComplete(): Unit = {
-    onCompleteOrError(null)
+  private def sendOnNextToAll(subscribers: Array[Subscriber[T]], elem: T): Future[Ack] = {
+    // counter that's only used when we go async, hence the null
+    var result: PromiseCounter[Continue.type] = null
+
+    var index = 0
+    while (index < subscribers.length) {
+      val subscriber = subscribers(index)
+      index += 1
+
+      // using the scheduler defined by each subscriber
+      import subscriber.scheduler
+
+      val ack =
+        try subscriber.onNext(elem)
+        catch { case NonFatal(ex) => Future.failed(ex) }
+
+      // if execution is synchronous, takes the fast-path
+      if (ack.isCompleted) {
+        // subscriber canceled or triggered an error? Then remove!
+        if (ack != Continue && ack.value.get != Continue.AsSuccess)
+          unsubscribe(subscriber)
+      }
+      else {
+        // going async, so we've got to count active futures for final Ack
+        // the counter starts from 1 because zero implies isCompleted
+        if (result == null) result = PromiseCounter(Continue, 1)
+        result.acquire()
+
+        ack.onComplete {
+          case Continue.AsSuccess =>
+            result.countdown()
+          case _ =>
+            // subscriber canceled or triggered an error? then remove
+            unsubscribe(subscriber)
+            result.countdown()
+        }
+      }
+    }
+
+    // has fast-path for completely synchronous invocation
+    if (result == null) Continue else {
+      result.countdown()
+      result.future
+    }
   }
 
   @tailrec
-  private def onCompleteOrError(ex: Throwable): Unit = {
+  private def sendOnCompleteOrError(ex: Throwable): Unit = {
     val state = stateRef.get
-    val subscribers = state.subscribers
-    val isDone = subscribers eq null
+    val set = state.subscribers
+    val subscribers: Iterable[Subscriber[T]] =
+      if (state.cache ne null) state.cache.toSeq else set
 
-    if (!isDone) {
-      // because of this CAS operation we are guaranteed to observe
+    if (subscribers ne null) {
+      // Because of this CAS operation we are guaranteed to observe
       // the most recent set of subscribers that may contain references
       // that haven't been seen in onNext yet
       if (!stateRef.compareAndSet(state, state.complete(ex)))
-        onCompleteOrError(ex)
+        sendOnCompleteOrError(ex)
       else {
-        val iterator = subscribers.iterator
+        val iterator = set.iterator
         while (iterator.hasNext) {
           val ref = iterator.next()
           if (ex != null)
@@ -194,15 +207,18 @@ object PublishSubject {
     */
   private[subjects] final case class State[T](
     subscribers: Set[Subscriber[T]] = Set.empty[Subscriber[T]],
+    cache: Array[Subscriber[T]] = null,
     errorThrown: Throwable = null) {
 
-    def isDone: Boolean = {
+    def refresh: State[T] =
+      copy(cache = subscribers.toArray)
+
+    def isDone: Boolean =
       subscribers eq null
-    }
 
     def complete(errorThrown: Throwable): State[T] = {
       if (subscribers eq null) this else
-        State[T](null, errorThrown)
+        State[T](null, null, errorThrown)
     }
   }
 }